Crate ic_embedders

Re-exports

pub use wasmtime_embedder::WasmtimeEmbedder;

pub use wasmtime_embedder::WasmtimeMemoryCreator;

Modules

cow_memory_creator

wasmtime_embedder

Structs

ExecSelect

ExtraWorkerHandle

InstanceRunResult

ProcessingGroup

QueueConfig

ResumeToken

ReturnToken

RunnerConfig

RunnerInput

RunnerOutput

WasmExecutionInput

WasmExecutionOutput

WasmExecutionResult

Enums

ExecutionResult

Traits

Embedder

ICMemoryCreator

Instance
Instance trait must be implemented by the wrapper of a concrete Wasm executer to hide all it's dependencies and internal details of its instance.

LinearMemory

Functions

spawn_new_runner