Trait ic_embedders::Instance

pub trait Instance {
    fn run(
        &mut self, 
        api: &mut (dyn SystemApi + 'static), 
        func_ref: FuncRef
    ) -> HypervisorResult<InstanceRunResult>;

    fn set_num_instructions(&mut self, cycles: NumInstructions);

    fn get_num_instructions(&self) -> NumInstructions;

    fn heap_size(&self) -> NumWasmPages;

    fn get_exported_globals(&self) -> Vec<Global>;

    unsafe fn heap_addr(&self) -> *const u8;

    fn get_stats(&self) -> InstanceStats;
}

Instance trait must be implemented by the wrapper of a concrete Wasm executer to hide all it's dependencies and internal details of its instance.

Required methods

fn run(
&mut self,
api: &mut (dyn SystemApi + 'static),
func_ref: FuncRef
) -> HypervisorResult<InstanceRunResult>

Executes first exported method on an embedder instance, whose name consists of one of the prefixes and method_name.

fn set_num_instructions(&mut self, cycles: NumInstructions)

Sets the amount of cycles for a method execution.

fn get_num_instructions(&self) -> NumInstructions

Returns the amount of cycles left.

fn heap_size(&self) -> NumWasmPages

Return the heap size.

fn get_exported_globals(&self) -> Vec<Global>

unsafe fn heap_addr(&self) -> *const u8

Return the heap address. If the Instance does not contain any memory, the pointer is null.

Safety

This function returns a pointer to Instance's memory. The pointer is only valid while the Instance object is kept alive.

fn get_stats(&self) -> InstanceStats

Returns execution statistics for this instance. Note that stats must be available even if this instance trapped.

Loading content...